Remove the extension-client and default loader for Extensions
