--- clients/extension/.npmignore
@@ -1,7 +1,0 @@
-# start off ignoring everything, and then add back only the files we want
-*
-!lib/**/*.d.ts
-!lib/**/*.d.ts.map
-!lib/**/*.js
-!lib/**/*.js.map
-test
