--- clients/extension/CHANGELOG.md
@@ -1,491 +1,0 @@
-# Change Log - @bentley/extension-client
-
-This log was last generated on Fri, 20 Aug 2021 17:47:22 GMT and should not be manually modified.
-
-## 2.19.7
-Fri, 20 Aug 2021 17:47:22 GMT
-
-_Version update only_
-
-## 2.19.6
-Tue, 17 Aug 2021 20:34:28 GMT
-
-_Version update only_
-
-## 2.19.5
-Fri, 13 Aug 2021 21:48:08 GMT
-
-_Version update only_
-
-## 2.19.4
-Thu, 12 Aug 2021 13:09:26 GMT
-
-_Version update only_
-
-## 2.19.3
-Wed, 04 Aug 2021 20:29:34 GMT
-
-_Version update only_
-
-## 2.19.2
-Tue, 03 Aug 2021 18:26:23 GMT
-
-_Version update only_
-
-## 2.19.1
-Thu, 29 Jul 2021 20:01:11 GMT
-
-_Version update only_
-
-## 2.19.0
-Mon, 26 Jul 2021 12:21:25 GMT
-
-_Version update only_
-
-## 2.18.4
-Tue, 10 Aug 2021 19:35:13 GMT
-
-_Version update only_
-
-## 2.18.3
-Wed, 28 Jul 2021 17:16:30 GMT
-
-_Version update only_
-
-## 2.18.2
-Mon, 26 Jul 2021 16:18:31 GMT
-
-_Version update only_
-
-## 2.18.1
-Fri, 16 Jul 2021 17:45:09 GMT
-
-_Version update only_
-
-## 2.18.0
-Fri, 09 Jul 2021 18:11:24 GMT
-
-_Version update only_
-
-## 2.17.3
-Mon, 26 Jul 2021 16:08:36 GMT
-
-_Version update only_
-
-## 2.17.2
-Thu, 08 Jul 2021 15:23:00 GMT
-
-_Version update only_
-
-## 2.17.1
-Fri, 02 Jul 2021 15:38:30 GMT
-
-_Version update only_
-
-## 2.17.0
-Mon, 28 Jun 2021 16:20:11 GMT
-
-_Version update only_
-
-## 2.16.10
-Thu, 22 Jul 2021 20:23:45 GMT
-
-_Version update only_
-
-## 2.16.9
-Tue, 06 Jul 2021 22:08:34 GMT
-
-_Version update only_
-
-## 2.16.8
-Fri, 02 Jul 2021 17:40:46 GMT
-
-_Version update only_
-
-## 2.16.7
-Mon, 28 Jun 2021 18:13:04 GMT
-
-_Version update only_
-
-## 2.16.6
-Mon, 28 Jun 2021 13:12:55 GMT
-
-_Version update only_
-
-## 2.16.5
-Fri, 25 Jun 2021 16:03:01 GMT
-
-_Version update only_
-
-## 2.16.4
-Wed, 23 Jun 2021 17:09:07 GMT
-
-_Version update only_
-
-## 2.16.3
-Wed, 16 Jun 2021 20:29:32 GMT
-
-_Version update only_
-
-## 2.16.2
-Thu, 03 Jun 2021 18:08:11 GMT
-
-_Version update only_
-
-## 2.16.1
-Thu, 27 May 2021 20:04:22 GMT
-
-_Version update only_
-
-## 2.16.0
-Mon, 24 May 2021 15:58:39 GMT
-
-_Version update only_
-
-## 2.15.6
-Wed, 26 May 2021 15:55:18 GMT
-
-_Version update only_
-
-## 2.15.5
-Thu, 20 May 2021 15:06:26 GMT
-
-_Version update only_
-
-## 2.15.4
-Tue, 18 May 2021 21:59:07 GMT
-
-_Version update only_
-
-## 2.15.3
-Mon, 17 May 2021 13:31:37 GMT
-
-_Version update only_
-
-## 2.15.2
-Wed, 12 May 2021 18:08:13 GMT
-
-_Version update only_
-
-## 2.15.1
-Wed, 05 May 2021 13:18:31 GMT
-
-_Version update only_
-
-## 2.15.0
-Fri, 30 Apr 2021 12:36:58 GMT
-
-### Updates
-
-- Update service URL
-
-## 2.14.4
-Thu, 22 Apr 2021 21:07:33 GMT
-
-_Version update only_
-
-## 2.14.3
-Thu, 15 Apr 2021 15:13:16 GMT
-
-_Version update only_
-
-## 2.14.2
-Thu, 08 Apr 2021 14:30:09 GMT
-
-_Version update only_
-
-## 2.14.1
-Mon, 05 Apr 2021 16:28:00 GMT
-
-_Version update only_
-
-## 2.14.0
-Fri, 02 Apr 2021 13:18:42 GMT
-
-_Version update only_
-
-## 2.13.0
-Tue, 09 Mar 2021 20:28:13 GMT
-
-### Updates
-
-- Updated to use TypeScript 4.1
-- begin rename project from iModel.js to iTwin.js
-
-## 2.12.3
-Mon, 08 Mar 2021 15:32:00 GMT
-
-_Version update only_
-
-## 2.12.2
-Wed, 03 Mar 2021 18:48:52 GMT
-
-_Version update only_
-
-## 2.12.1
-Tue, 23 Feb 2021 20:54:45 GMT
-
-_Version update only_
-
-## 2.12.0
-Thu, 18 Feb 2021 22:10:12 GMT
-
-_Version update only_
-
-## 2.11.2
-Thu, 18 Feb 2021 02:50:59 GMT
-
-_Version update only_
-
-## 2.11.1
-Thu, 04 Feb 2021 17:22:41 GMT
-
-_Version update only_
-
-## 2.11.0
-Thu, 28 Jan 2021 13:39:27 GMT
-
-_Version update only_
-
-## 2.10.3
-Fri, 08 Jan 2021 18:34:03 GMT
-
-_Version update only_
-
-## 2.10.2
-Fri, 08 Jan 2021 14:52:02 GMT
-
-_Version update only_
-
-## 2.10.1
-Tue, 22 Dec 2020 00:53:38 GMT
-
-_Version update only_
-
-## 2.10.0
-Fri, 18 Dec 2020 18:24:01 GMT
-
-_Version update only_
-
-## 2.9.9
-Sun, 13 Dec 2020 19:00:03 GMT
-
-_Version update only_
-
-## 2.9.8
-Fri, 11 Dec 2020 02:57:36 GMT
-
-_Version update only_
-
-## 2.9.7
-Wed, 09 Dec 2020 20:58:23 GMT
-
-_Version update only_
-
-## 2.9.6
-Mon, 07 Dec 2020 18:40:48 GMT
-
-_Version update only_
-
-## 2.9.5
-Sat, 05 Dec 2020 01:55:56 GMT
-
-_Version update only_
-
-## 2.9.4
-Wed, 02 Dec 2020 20:55:40 GMT
-
-_Version update only_
-
-## 2.9.3
-Mon, 23 Nov 2020 20:57:56 GMT
-
-_Version update only_
-
-## 2.9.2
-Mon, 23 Nov 2020 15:33:50 GMT
-
-_Version update only_
-
-## 2.9.1
-Thu, 19 Nov 2020 17:03:42 GMT
-
-_Version update only_
-
-## 2.9.0
-Wed, 18 Nov 2020 16:01:50 GMT
-
-_Version update only_
-
-## 2.8.1
-Tue, 03 Nov 2020 00:33:56 GMT
-
-_Version update only_
-
-## 2.8.0
-Fri, 23 Oct 2020 17:04:02 GMT
-
-_Version update only_
-
-## 2.7.6
-Wed, 11 Nov 2020 16:28:23 GMT
-
-_Version update only_
-
-## 2.7.5
-Fri, 23 Oct 2020 16:23:50 GMT
-
-_Version update only_
-
-## 2.7.4
-Mon, 19 Oct 2020 17:57:01 GMT
-
-_Version update only_
-
-## 2.7.3
-Wed, 14 Oct 2020 17:00:58 GMT
-
-_Version update only_
-
-## 2.7.2
-Tue, 13 Oct 2020 18:20:38 GMT
-
-_Version update only_
-
-## 2.7.1
-Thu, 08 Oct 2020 13:04:35 GMT
-
-_Version update only_
-
-## 2.7.0
-Fri, 02 Oct 2020 18:03:32 GMT
-
-### Updates
-
-- Fixed lint warnings.
-
-## 2.6.5
-Sat, 26 Sep 2020 16:06:34 GMT
-
-_Version update only_
-
-## 2.6.4
-Tue, 22 Sep 2020 17:40:07 GMT
-
-_Version update only_
-
-## 2.6.3
-Mon, 21 Sep 2020 14:47:09 GMT
-
-_Version update only_
-
-## 2.6.2
-Mon, 21 Sep 2020 13:07:43 GMT
-
-_Version update only_
-
-## 2.6.1
-Fri, 18 Sep 2020 13:15:08 GMT
-
-_Version update only_
-
-## 2.6.0
-Thu, 17 Sep 2020 13:16:12 GMT
-
-### Updates
-
-- Moved ESLint configuration to a plugin
-
-## 2.5.5
-Wed, 02 Sep 2020 17:42:23 GMT
-
-_Version update only_
-
-## 2.5.4
-Fri, 28 Aug 2020 15:34:15 GMT
-
-_Version update only_
-
-## 2.5.3
-Wed, 26 Aug 2020 11:46:00 GMT
-
-_Version update only_
-
-## 2.5.2
-Tue, 25 Aug 2020 22:09:08 GMT
-
-_Version update only_
-
-## 2.5.1
-Mon, 24 Aug 2020 18:13:04 GMT
-
-_Version update only_
-
-## 2.5.0
-Thu, 20 Aug 2020 20:57:09 GMT
-
-### Updates
-
-- Switch to ESLint
-
-## 2.4.2
-Fri, 14 Aug 2020 16:34:09 GMT
-
-_Version update only_
-
-## 2.4.1
-Fri, 07 Aug 2020 19:57:43 GMT
-
-_Version update only_
-
-## 2.4.0
-Tue, 28 Jul 2020 16:26:24 GMT
-
-_Version update only_
-
-## 2.3.3
-Thu, 23 Jul 2020 12:57:15 GMT
-
-_Version update only_
-
-## 2.3.2
-Tue, 14 Jul 2020 23:50:36 GMT
-
-_Version update only_
-
-## 2.3.1
-Mon, 13 Jul 2020 18:50:13 GMT
-
-_Version update only_
-
-## 2.3.0
-Fri, 10 Jul 2020 17:23:14 GMT
-
-_Version update only_
-
-## 2.2.1
-Tue, 07 Jul 2020 14:44:52 GMT
-
-_Version update only_
-
-## 2.2.0
-Fri, 19 Jun 2020 14:10:03 GMT
-
-_Version update only_
-
-## 2.1.0
-Thu, 28 May 2020 22:48:59 GMT
-
-_Version update only_
-
-## 2.0.0
-Wed, 06 May 2020 13:17:49 GMT
-
-### Updates
-
-- Update for new delete extension API
-- Initial version
-- order imports.
-- react to creation of context-registry-client
-- Upgrade to Rush 5.23.2
-
