--- clients/extension/README.md
@@ -1,12 +1,0 @@
-# iTwin.js Extension Client
-
-Copyright © Bentley Systems, Incorporated. All rights reserved. See LICENSE.md for license terms and full copyright notice.
-
-## Description
-
-The __@bentley/extension-client__ package contains client wrappers for querying, uploading and deleting iTwin.js Extensions from the Extension Service.
-
-## Documentation
-
-See the [iTwin.js](https://www.itwinjs.org) documentation for more information.
-
