--- clients/extension/certa.json
@@ -1,28 +1,0 @@
-{
-  // Comments are allowed here!
-  "testBundle": "./lib/test/webpack/bundled-tests.js",
-  "instrumentedTestBundle": "./lib/test/webpack/bundled-tests.instrumented.js",
-  "backendInitModule": "./node_modules/@bentley/oidc-signin-tool/lib/certa/certaBackend.js",
-  "ports": {
-    "debugging": 5858,
-    "frontend": 3000,
-    "frontendDebugging": 9223
-  },
-  "mochaOptions": {
-    "timeout": 50000,
-    "reporter": "node_modules/@bentley/build-tools/mocha-reporter",
-    "reporterOptions": {
-      "mochaFile": "lib/test/junit_results.xml"
-    }
-  },
-  "chromeOptions": {
-    "publicDirs": [
-      "src/public",
-      "src/test/public"
-    ],
-    "args": [
-      "--disable-web-security",
-      "--no-sandbox"
-    ]
-  }
-}
