--- clients/extension/package.json
@@ -1,69 +1,0 @@
-{
-  "name": "@bentley/extension-client",
-  "version": "3.0.0-dev.34",
-  "description": "IModel.js Extension Client",
-  "main": "lib/extension-client.js",
-  "typings": "lib/extension-client",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/imodeljs/imodeljs/tree/master/clients/extension"
-  },
-  "scripts": {
-    "compile": "npm run build",
-    "build": "tsc 1>&2",
-    "clean": "rimraf lib .rush/temp/package-deps*.json",
-    "cover": "",
-    "docs": "betools docs --includes=../../generated-docs/extract --json=../../generated-docs/clients/extension-client/file.json --tsIndexFile=./extension-client.ts --onlyJson",
-    "extract-api": "betools extract-api --entry=extension-client",
-    "lint": "eslint -f visualstudio \"./src/**/*.ts\" 1>&2",
-    "test": "",
-    "test:integration": "npm run webpackTests && npm run test:integration:chrome",
-    "test:integration:chrome": "certa -r chrome",
-    "test:integration:electron": "certa -r electron",
-    "test:integration:node": "certa -r node",
-    "copy:config": "internal-tools copy-config",
-    "webpackTests": "webpack --config ./src/test/utils/webpack.config.js 1>&2"
-  },
-  "keywords": [
-    "Bentley",
-    "BIM",
-    "iModel",
-    "Extension"
-  ],
-  "author": {
-    "name": "Bentley Systems, Inc.",
-    "url": "http://www.bentley.com"
-  },
-  "dependencies": {
-    "@bentley/bentleyjs-core": "workspace:*",
-    "@bentley/imodeljs-common": "workspace:*",
-    "@bentley/itwin-client": "workspace:*"
-  },
-  "devDependencies": {
-    "@bentley/build-tools": "workspace:*",
-    "@bentley/certa": "workspace:*",
-    "@bentley/context-registry-client": "workspace:*",
-    "@bentley/eslint-plugin": "workspace:*",
-    "@bentley/oidc-signin-tool": "workspace:*",
-    "@types/chai": "^4.1.4",
-    "@types/mocha": "^8.2.2",
-    "@types/node": "10.14.1",
-    "chai": "^4.1.2",
-    "dotenv": "^10.0.0",
-    "dotenv-expand": "^5.1.0",
-    "eslint": "^7.11.0",
-    "internal-tools": "workspace:*",
-    "mocha": "^8.3.2",
-    "rimraf": "^3.0.2",
-    "source-map-loader": "^1.0.0",
-    "typescript": "~4.3.0",
-    "webpack": "4.42.0"
-  },
-  "eslintConfig": {
-    "plugins": [
-      "@bentley"
-    ],
-    "extends": "plugin:@bentley/imodeljs-recommended"
-  }
-}
