--- clients/extension/src/extension-client.ts
@@ -1,16 +1,0 @@
-/*---------------------------------------------------------------------------------------------
-* Copyright (c) Bentley Systems, Incorporated. All rights reserved.
-* See LICENSE.md in the project root for license terms and full copyright notice.
-*--------------------------------------------------------------------------------------------*/
-export * from "./ExtensionClient";
-export * from "./ExtensionFile";
-export * from "./ExtensionProps";
-
-/** @docs-package-description
- * The extension-client package provides a set of APIs to load and manage iModel.js Extensions.
- */
-
-/**
- * @docs-group-description ExtensionService
- * Classes for communicating with an iModel.js Extension Service.
- */
