--- clients/extension/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./node_modules/@bentley/build-tools/tsconfig-base.json",
-  "compilerOptions": {
-    "target": "es2017",
-    "module": "commonjs",
-    "outDir": "./lib"
-  },
-  "include": [
-    "./src/**/*.ts"
-  ],
-  "exclude": [
-    "lib",
-    "node_modules"
-  ]
-}
