--- common/api/extension-client.api.md
@@ -1,60 +1,0 @@
-## API Report File for "@bentley/extension-client"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-
-import { AuthorizedClientRequestContext } from '@bentley/itwin-client';
-import { Client } from '@bentley/itwin-client';
-import { ClientRequestContext } from '@bentley/bentleyjs-core';
-
-// @beta
-export class ExtensionClient extends Client {
-    constructor();
-    // @internal
-    createExtension(requestContext: AuthorizedClientRequestContext, contextId: string, extensionName: string, version: string, checksum: string, file: Buffer): Promise<void>;
-    deleteExtension(requestContext: AuthorizedClientRequestContext, contextId: string, extensionName: string, version?: string): Promise<void>;
-    downloadExtension(requestContext: AuthorizedClientRequestContext, contextId: string, extensionName: string, version: string): Promise<ExtensionFile[]>;
-    getExtensionProps(requestContext: AuthorizedClientRequestContext, contextId: string, extensionName: string, version: string): Promise<ExtensionProps | undefined>;
-    getExtensions(requestContext: AuthorizedClientRequestContext, contextId: string, extensionName?: string): Promise<ExtensionProps[]>;
-    // (undocumented)
-    getUrl(requestContext: ClientRequestContext): Promise<string>;
-    // @internal (undocumented)
-    protected getUrlSearchKey(): string;
-    }
-
-// @beta
-export interface ExtensionFile {
-    // (undocumented)
-    content: ArrayBuffer;
-    // (undocumented)
-    fileName: string;
-}
-
-// @beta
-export interface ExtensionProps {
-    // (undocumented)
-    contextId: string;
-    // (undocumented)
-    extensionName: string;
-    // (undocumented)
-    files: FileInfo[];
-    // (undocumented)
-    isPublic: boolean;
-    // (undocumented)
-    status: ExtensionUploadStatus;
-    // (undocumented)
-    timestamp: Date;
-    // (undocumented)
-    uploadedBy: string;
-    // (undocumented)
-    version: string;
-}
-
-// @internal
-export function extensionPropsFromJSON(jsonObject: any): ExtensionProps | undefined;
-
-
-// (No @packageDocumentation comment for this package)
-
-```
