--- common/changes/@bentley/extension-client/2.19.1-changelogs_2021-07-29-20-10.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "comment": "",
-      "type": "none",
-      "packageName": "@bentley/extension-client"
-    }
-  ],
-  "packageName": "@bentley/extension-client",
-  "email": "32379572+skirby1996@users.noreply.github.com"
-}
