--- common/changes/@bentley/extension-client/2.19.2-changelogs_2021-08-03-18-30.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "comment": "",
-      "type": "none",
-      "packageName": "@bentley/extension-client"
-    }
-  ],
-  "packageName": "@bentley/extension-client",
-  "email": "32379572+skirby1996@users.noreply.github.com"
-}
