--- common/changes/@bentley/extension-client/2.19.5-changelogs_2021-08-13-21-56.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "comment": "",
-      "type": "none",
-      "packageName": "@bentley/extension-client"
-    }
-  ],
-  "packageName": "@bentley/extension-client",
-  "email": "32379572+skirby1996@users.noreply.github.com"
-}
