--- common/changes/@bentley/extension-client/spencer-bump-dotenv-dependency_2021-08-03-18-06.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "packageName": "@bentley/extension-client",
-      "comment": "Bumped dotenv and removed deprecated @types/dotenv",
-      "type": "none"
-    }
-  ],
-  "packageName": "@bentley/extension-client",
-  "email": "54333460+SpencerWBarnes@users.noreply.github.com"
-}
