--- common/changes/@bentley/extension-client/spencer-remove-deprecated-configloader_2021-07-28-20-49.json
@@ -1,11 +1,0 @@
-{
-  "changes": [
-    {
-      "packageName": "@bentley/extension-client",
-      "comment": "Removed references to the deleted config-loader package",
-      "type": "none"
-    }
-  ],
-  "packageName": "@bentley/extension-client",
-  "email": "54333460+SpencerWBarnes@users.noreply.github.com"
-}
