+++ common/changes/@bentley/frontend-devtools/remove-extension-client-dep_2021-08-23-20-58.json
@@ -1,0 +1,10 @@
+{
+  "changes": [
+    {
+      "packageName": "@bentley/frontend-devtools",
+      "comment": "",
+      "type": "none"
+    }
+  ],
+  "packageName": "@bentley/frontend-devtools"
+}
