+++ core/frontend/package.json
@@ -96,7 +96,6 @@
   ],
   "dependencies": {
     "@bentley/context-registry-client": "workspace:*",
-    "@bentley/extension-client": "workspace:*",
     "@bentley/reality-data-client": "workspace:*",
     "fuse.js": "^3.3.0",
     "semver": "^5.5.0",
